Split knockout binding handlers into separate file for safe importing
Fixes the hatchery filtering and various other issues from load order breaking extenders.
