+++ src/modules/index.ts
@@ -4,5 +4,8 @@ import './polyfill';
 // Ensure that the Knockout Extenders are injected
 import './koExtenders';
 
+// Inject the Knockout bindingHandlers (may load other local modules due to imports)
+import './koBindingHandlers';
+
 // Load everything else
 import './temporaryWindowInjection';
