+++ src/modules/koBindingHandlers.d.ts
@@ -1,0 +1,8 @@
+export declare module 'knockout' {
+    export interface BindingHandlers {
+        contentEditable;
+        playerSpriteMove;
+        sortable;
+        resizable;
+    }
+}
