+++ src/modules/koExtenders.ts
@@ -1,9 +1,18 @@
-/* eslint-disable no-param-reassign */
+// /* eslint-disable no-param-reassign */
 /// <reference path="./koExtenders.d.ts" />
 
-import Sortable from 'sortablejs';
 import type { Subscribable, Observable, Computed } from 'knockout';
-import Settings from './settings';
+
+/*
+    WARNING: Avoid importing local modules here at all costs!
+
+    This file absolutely must be loaded before these extenders can be used elsewhere. The compiler cannot
+    detect dependencies on these extenders on its own and attempts to use them before this file loads
+    will fail silently. Local imports here would make the evaluation order unpredictable and unstable;
+    minor commits could break unrelated parts of the game. This is a nightmare to debug! Don't do it!
+
+    When using these extenders in a module, import this file to ensure proper dependency tracking.
+*/
 
 // Only numeric values allowed - usage: ko.observable(0).extend({ numeric: 0 });
 ko.extenders.numeric = (target: Subscribable, precision: number) => {
@@ -108,154 +117,4 @@ ko.extenders.skippableRateLimit = (target: Subscribable, delay: number) => {
         return startRateLimit;
     };
     return target.extend({ rateLimit: { timeout: delay, method: skippableLimiter } }) as typeof target & SkippableRateLimit;
-};
-
-ko.bindingHandlers.contentEditable = {
-    init: (element: HTMLElement, valueAccessor) => {
-        const value = valueAccessor();
-
-        function onBlur() {
-            if (ko.isWriteableObservable(value)) {
-                value(this.textContent);
-            }
-        }
-
-        element.textContent = value();
-        element.contentEditable = 'true';
-        element.addEventListener('blur', onBlur);
-    },
-};
-
-//Fixes the PlayerSVG rendering beneath other SVGs by rerendering the current PlayerSVG in the forefront. #4306
-//TODO: Replace with logic that maintains a singular PlayerSpriteSVG (Might be performance costly over a simple observer)
-function handleVisibleElement(element) {
-    if (element.classList.contains('iconLocation')) {
-        if (element.classList.contains('iconLocation')) {
-            var targetElement = document.getElementById('playerSprite');
-            var imageElement = element.cloneNode(true);
-            imageElement.classList.remove('hide');
-
-            targetElement.innerHTML = '';
-            targetElement.appendChild(imageElement);
-
-            //Get required variables to replicate the rotate on parent group as well
-            var rotate = imageElement.getAttribute('rotate');
-            var x = imageElement.getAttribute('localx');
-            var y = imageElement.getAttribute('localy');
-
-            if (rotate) {
-                targetElement.setAttribute('transform', 'rotate(90,' + x + ', ' + y + ')');
-            } else {
-                targetElement.setAttribute('transform', '');
-            }
-        }
-    }
-}
-
-ko.bindingHandlers.playerSpriteMove = {
-    init: function (element) {
-        const observer = new MutationObserver((mutationsList) => {
-            for (const mutation of mutationsList) {
-                if (mutation.type === 'attributes' && mutation.attributeName === 'style') {
-                    const target = mutation.target as HTMLElement;
-                    const displayStyle = window.getComputedStyle(target).getPropertyValue('display');
-
-                    const isVisible = displayStyle !== 'none';
-                    if (isVisible) {
-                        handleVisibleElement(element);
-                    }
-                }
-            }
-        });
-
-        // Observe changes to the 'style' attribute of the target element
-        observer.observe(element, { attributes: true });
-
-        // Trigger initial visibility check on load
-        const displayStyle = window.getComputedStyle(element).getPropertyValue('display');
-        const isVisible = displayStyle !== 'none';
-        if (isVisible) {
-            handleVisibleElement(element);
-        }
-    },
-};
-//TODO END
-
-const sortableControllers = new WeakMap();
-ko.bindingHandlers.sortable = {
-    init: function (element, valueAccessor, allBindings, viewModel) {
-        sortableControllers.set(element, null);
-        const value = valueAccessor();
-        ko.applyBindingsToNode(element, {
-            template: {
-                ...value,
-                afterRender(nodes, el) {
-                    nodes.forEach((n) => {
-                        if (n.dataset) {
-                            n.dataset.sortableId = value.options.getId(el);
-                        }
-                    });
-                },
-                beforeRemove(node: HTMLElement, idx, el) {
-                    // Sortable may have cloned the node, so we need to get the real one
-                    const found = element.querySelector(`[data-sortable-id="${value.options.getId(el)}"]`);
-                    (found ?? node)?.remove();
-                },
-            },
-        }, viewModel);
-        return { controlsDescendantBindings: true };
-    },
-    update: function (element, valueAccessor) {
-        const value = ko.unwrap(valueAccessor());
-        const options = {
-            // defaults
-            animation: 100,
-            sort: true,
-            delay: 500,
-            delayOnTouchOnly: true,
-            touchStartThreshold: 20,
-
-            // override with options passed through knockout binding
-            ...(value.options ?? {}),
-            dataIdAttr: 'data-sortable-id',
-
-            // handle updating underlying knockout array when moving items
-            onEnd: (evt, originalEvt) => {
-                value.options?.onEnd?.(evt, originalEvt);
-
-                const { oldIndex, newIndex } = evt;
-                const list = [...value.foreach()];
-                const movedItem = list.splice(oldIndex, 1)[0];
-
-                const newList = [
-                    ...list.slice(0, newIndex),
-                    movedItem,
-                    ...list.slice(newIndex),
-                ];
-                value.foreach(newList);
-            },
-        };
-
-        sortableControllers.set(element, Sortable.create(element, options));
-    },
-};
-
-const moduleResizeObserver = new ResizeObserver((entries) => {
-    entries.forEach((entry) => {
-        const height = entry.target.getBoundingClientRect().height;
-        if (height > 0) {
-            const resizeId = (entry.target as HTMLElement).dataset.resizeId;
-            Settings.setSettingByName(`moduleHeight.${resizeId}`, height);
-        }
-    });
-});
-
-ko.bindingHandlers.resizable = {
-    init: function (element, valueAccessor) {
-        const value = valueAccessor();
-        element.classList.add('resizable-container');
-        element.style.height = `${Settings.getSetting(`moduleHeight.${value.setting}`).value}px`;
-        element.dataset.resizeId = value.setting;
-        moduleResizeObserver.observe(element);
-    },
 };
